:sparkles: Create blinky

+++ firmware/pico2/src/common/reg.rs
@@ -1,8 +1,8 @@
 #[repr(usize)]
 #[derive(Clone, Copy)]
 pub enum RegAddr {
-    RESET = 0x4002_0000,
-    IO_BANK0 = 0x4002_8000,
-    SIO = 0xd000_0000,
-    PADS_BANK0 = 0x4003_8000,
+    RESET = 0x4002_0000 as usize,
+    IO_BANK0 = 0x4002_8000 as usize,
+    SIO = 0xd000_0000 as usize,
+    PADS_BANK0 = 0x4003_8000 as usize,
 }
